// src/pages/DetailsCardPage/types.ts
export type DetailsCard = {
  format: string;
  description: string;
  price: string;
  artist: string;
  album: string;
  diameter: number;
  rpm: number;
  genre: string;
  style: string;
  recordLabel: string;
  weight: number;
  condition: string;
  coverCondition: string;
  images: string;
  id: string;
  releaseYear: number;
};